% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{downloadData}
\alias{downloadData}
\title{Interface with Rocket-API}
\usage{
downloadData(id)
}
\arguments{
\item{id}{An object to save.}
}
\description{
pigz needs to be installed for this function to work.
}
